Multi-Threading, Factory Fix, Intersection interface....
Multi threading was not working because factory returned the same object every time.
now it returns a new object. I also added a intersection interfaces but the methods inside needs to be adjust to be able to work with other intersection typs. i also added a name to the traffic light so the console print is now more useful. also an Intersection Factaory is now avadible